requirements.txt
# Core
openai>=1.30.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# PDF processing
PyMuPDF>=1.24.0

# LangGraph agent
langgraph>=0.2.0

# LlamaIndex pipeline
llama-index-core>=0.11.0
llama-index-llms-openai>=0.2.0
llama-index-embeddings-openai>=0.2.0
llama-index-readers-file>=0.2.0

# Evaluation & visualisation
matplotlib>=3.7.0

# Streamlit app
streamlit>=1.35.0
sentence-transformers

tiktoken>=0.7.0
